新增品牌关联的分类

+++ gulimall-product/src/main/java/cn/wjk/gulimall/product/domain/dto/CategoryBrandRelationDTO.java
@@ -12,6 +12,6 @@ import lombok.Data;
  */
 @Data
 public class CategoryBrandRelationDTO {
-    private Integer brandId;
-    private Integer catelogId;
+    private Long brandId;
+    private Long catelogId;
 }
